+++ runtime/src/lib.rs
@@ -138,9 +138,9 @@ pub type Executive = frame_executive::Executive<
 ///     charged.
 pub struct WeightToFee;
 
-const P_FACTOR: u128 = 10;
-const Q_FACTOR: u128 = 100;
-const POLY_DEGREE: u8 = 1;
+pub const P_FACTOR: u128 = 10;
+pub const Q_FACTOR: u128 = 100;
+pub const POLY_DEGREE: u8 = 1;
 
 impl WeightToFeePolynomial for WeightToFee {
     type Balance = Balance;
@@ -219,28 +219,28 @@ pub const DAYS: BlockNumber = HOURS * 24;
 
 /// We assume that ~5% of the block weight is consumed by `on_initialize`
 /// handlers. This is used to limit the maximal weight of a single extrinsic.
-const AVERAGE_ON_INITIALIZE_RATIO: Perbill = Perbill::from_percent(5);
+pub const AVERAGE_ON_INITIALIZE_RATIO: Perbill = Perbill::from_percent(5);
 
 /// We allow `Normal` extrinsics to fill up the block up to 75%, the rest can be
 /// used by `Operational` extrinsics.
-const NORMAL_DISPATCH_RATIO: Perbill = Perbill::from_percent(75);
+pub const NORMAL_DISPATCH_RATIO: Perbill = Perbill::from_percent(75);
 
 /// We allow for 0.5 of a second of compute with a 12 second average block time.
-const MAXIMUM_BLOCK_WEIGHT: Weight = Weight::from_parts(
+pub const MAXIMUM_BLOCK_WEIGHT: Weight = Weight::from_parts(
     WEIGHT_REF_TIME_PER_SECOND.saturating_div(2),
     cumulus_primitives_core::relay_chain::MAX_POV_SIZE as u64,
 );
 
 /// Maximum number of blocks simultaneously accepted by the Runtime, not yet
 /// included into the relay chain.
-const UNINCLUDED_SEGMENT_CAPACITY: u32 = 1;
+pub const UNINCLUDED_SEGMENT_CAPACITY: u32 = 1;
 /// How many parachain blocks are processed by the relay chain per parent.
 /// Limits the number of blocks authored per slot.
-const BLOCK_PROCESSING_VELOCITY: u32 = 1;
+pub const BLOCK_PROCESSING_VELOCITY: u32 = 1;
 /// Relay chain slot duration, in milliseconds.
-const RELAY_CHAIN_SLOT_DURATION_MILLIS: u32 = 6000;
+pub const RELAY_CHAIN_SLOT_DURATION_MILLIS: u32 = 6000;
 /// Maximum length for a block.
-const MAX_BLOCK_LENGTH: u32 = 5 * 1024 * 1024;
+pub const MAX_BLOCK_LENGTH: u32 = 5 * 1024 * 1024;
 
 /// The version information used to identify this runtime when compiled
 /// natively.
@@ -946,38 +946,14 @@ cumulus_pallet_parachain_system::register_validate_block! {
     BlockExecutor = cumulus_pallet_aura_ext::BlockExecutor::<Runtime, Executive>,
 }
 
+// tests
 #[cfg(test)]
 mod tests {
-
-    use frame_support::traits::TypedGet;
-
     use super::*;
 
+    // RUNTIME_API_VERSIONS constant is generated by a macro and is private.
     #[test]
-    fn weight_to_fee_constants() {
-        assert_eq!(P_FACTOR, 10);
-
-        assert_eq!(Q_FACTOR, 100);
-
-        assert_eq!(POLY_DEGREE, 1);
-    }
-
-    #[test]
-    fn frame_system_constants() {
-        assert_eq!(
-            MAXIMUM_BLOCK_WEIGHT,
-            Weight::from_parts(
-                WEIGHT_REF_TIME_PER_SECOND.saturating_div(2),
-                cumulus_primitives_core::relay_chain::MAX_POV_SIZE as u64
-            )
-        );
-
-        assert_eq!(UNINCLUDED_SEGMENT_CAPACITY, 1);
-
-        assert_eq!(BLOCK_PROCESSING_VELOCITY, 1);
-
-        assert_eq!(RELAY_CHAIN_SLOT_DURATION_MILLIS, 6000);
-
+    fn check_version() {
         assert_eq!(
             VERSION,
             RuntimeVersion {
@@ -991,121 +967,5 @@ mod tests {
                 state_version: 1,
             }
         );
-
-        assert_eq!(MILLISECS_PER_BLOCK, 12000);
-
-        assert_eq!(SLOT_DURATION, MILLISECS_PER_BLOCK);
-
-        assert_eq!(MINUTES, 60_000 / (MILLISECS_PER_BLOCK as BlockNumber));
-
-        assert_eq!(HOURS, MINUTES * 60);
-
-        assert_eq!(DAYS, HOURS * 24);
-
-        assert_eq!(MAX_BLOCK_LENGTH, 5 * 1024 * 1024);
-
-        assert_eq!(SS58Prefix::get(), 42);
-
-        assert_eq!(<Runtime as frame_system::Config>::MaxConsumers::get(), 16);
-    }
-
-    #[test]
-    fn proxy_constants() {
-        assert_eq!(MaxProxies::get(), 32);
-
-        assert_eq!(MaxPending::get(), 32);
-
-        assert_eq!(ProxyDepositBase::get(), deposit(1, 40));
-
-        assert_eq!(AnnouncementDepositBase::get(), deposit(1, 48));
-
-        assert_eq!(ProxyDepositFactor::get(), deposit(0, 33));
-
-        assert_eq!(AnnouncementDepositFactor::get(), deposit(0, 66));
-    }
-
-    #[test]
-    fn balances_constants() {
-        assert_eq!(MaxFreezes::get(), 0);
-
-        assert_eq!(MaxHolds::get(), 0);
-
-        assert_eq!(MaxLocks::get(), 50);
-
-        assert_eq!(MaxReserves::get(), 50);
-    }
-
-    #[test]
-    fn assets_constants() {
-        assert_eq!(AssetDeposit::get(), 10 * CENTS);
-
-        assert_eq!(AssetAccountDeposit::get(), deposit(1, 16));
-
-        assert_eq!(ApprovalDeposit::get(), EXISTENTIAL_DEPOSIT);
-
-        assert_eq!(StringLimit::get(), 50);
-
-        assert_eq!(MetadataDepositBase::get(), deposit(1, 68));
-
-        assert_eq!(MetadataDepositPerByte::get(), deposit(0, 1));
-
-        assert_eq!(RemoveItemsLimit::get(), 1000);
-    }
-
-    #[test]
-    fn transaction_payment_constants() {
-        assert_eq!(TransactionByteFee::get(), 10 * MICROCENTS);
-
-        assert_eq!(OperationalFeeMultiplier::get(), 5);
-    }
-
-    #[test]
-    fn cumulus_pallet_parachain_system_constants() {
-        assert_eq!(ReservedXcmpWeight::get(), MAXIMUM_BLOCK_WEIGHT.saturating_div(4));
-
-        assert_eq!(ReservedDmpWeight::get(), MAXIMUM_BLOCK_WEIGHT.saturating_div(4));
-    }
-
-    #[test]
-    fn multisig_constants() {
-        assert_eq!(DepositBase::get(), deposit(1, 88));
-
-        assert_eq!(DepositFactor::get(), deposit(0, 32));
-
-        assert_eq!(MaxSignatories::get(), 100);
-    }
-
-    #[test]
-    fn session_constants() {
-        assert_eq!(Period::get(), 6 * HOURS);
-
-        assert_eq!(Offset::get(), 0);
-    }
-
-    #[test]
-    #[allow(clippy::assertions_on_constants)]
-    fn aura_constants() {
-        assert!(!AllowMultipleBlocksPerSlot::get());
-
-        assert_eq!(MaxAuthorities::get(), 100_000);
-    }
-
-    #[test]
-    fn collator_selection_constants() {
-        let pallet_id_to_string = |id: PalletId| -> String {
-            core::str::from_utf8(&id.0).unwrap_or_default().to_string()
-        };
-
-        assert_eq!(pallet_id_to_string(PotId::get()), pallet_id_to_string(PalletId(*b"PotStake")));
-
-        assert_eq!(SessionLength::get(), 6 * HOURS);
-
-        assert_eq!(StakingAdminBodyId::get(), BodyId::Defense);
-
-        assert_eq!(MaxCandidates::get(), 100);
-
-        assert_eq!(MaxInvulnerables::get(), 20);
-
-        assert_eq!(MinEligibleCollators::get(), 4);
     }
 }
